requirements.txt
asgiref==3.11.1
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
defusedxml==0.7.1
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
idna==3.13
mssql-django==1.7
oauthlib==3.3.1
pycparser==3.0
PyJWT==2.12.1
pyodbc==5.3.0
python3-openid==3.2.0
pytz==2026.1.post1
requests==2.33.1
requests-oauthlib==2.0.0
social-auth-app-django==5.8.0
social-auth-core==4.8.6
sqlparse==0.5.5
tzdata==2026.1
urllib3==2.6.3
openai==1.35.10
python-dotenv==0.21.0
